refactor(ui): decompose evaluation feature hooks and components
