+++ src/quodeq/ui/tools/size_grandfather.mjs
@@ -67,19 +67,13 @@ export const SIZE_GRANDFATHER = [
   "src/features/evaluation/components/BranchScopeSelector.jsx",
   "src/features/evaluation/components/CleanScanToggle.jsx",
   "src/features/evaluation/components/ConsoleLogViewer.jsx",
-  "src/features/evaluation/components/DimensionSelector.jsx",
-  "src/features/evaluation/components/EvaluationForm.jsx",
   "src/features/evaluation/components/EvaluationForm.test.jsx",
   "src/features/evaluation/components/EvaluationStatus.test.jsx",
   "src/features/evaluation/components/FolderBrowser.jsx",
   "src/features/evaluation/components/JobStatStrip.jsx",
   "src/features/evaluation/components/JobStatStrip.test.jsx",
-  "src/features/evaluation/components/LiveViolationsFeed.jsx",
   "src/features/evaluation/components/LiveViolationsFeed.test.jsx",
-  "src/features/evaluation/components/ReEvaluateCard.jsx",
   "src/features/evaluation/components/ReEvaluateCard.test.jsx",
-  "src/features/evaluation/components/ScanModeCards.jsx",
-  "src/features/evaluation/components/ScanProgress.jsx",
   "src/features/evaluation/components/ScanProgress.test.jsx",
   "src/features/evaluation/components/buildJobStatCells.js",
   "src/features/evaluation/components/buildJobStatCells.test.js",
@@ -90,7 +84,6 @@ export const SIZE_GRANDFATHER = [
   "src/features/evaluation/eval-log/useJobLogStream.js",
   "src/features/evaluation/eval-log/useJobLogStream.test.jsx",
   "src/features/evaluation/eval-log/useJobLogStream549.test.jsx",
-  "src/features/evaluation/hooks/useEvaluation.js",
   "src/features/evaluation/hooks/useEvaluation.test.jsx",
   "src/features/evaluation/hooks/usePluginDimensions.test.jsx",
   "src/features/evaluation/hooks/useRunEventStream.js",
